requirements.txt
streamlit>=1.28.0
tensorflow>=2.13.0
pillow>=10.0.0
numpy>=1.24.0 